Try only running exhaustive tests with latest Perl
Installation of some of the deeper analysis packages fail on older
versions of Perl, but we should only need to do things like coverage
and style testing on the latest Perl version.  Skip installing
dependencies on older versions to speed up the Travis-CI tests and
allow us to keep testing with really old versions of Perl.